+++ test/ksi_hashchain_test.c
@@ -100,7 +100,7 @@ cleanup:
 
 static void buildHashChain(CuTest *tc, const char *hexImprint, int isLeft, int levelCorrection, KSI_LIST(KSI_HashChainLink) **chn) {
 	unsigned char buf[1024];
-	unsigned buf_len;
+	size_t buf_len;
 	int res;
 	KSI_DataHash *hsh = NULL;
 
@@ -138,7 +138,7 @@ static void testCalChainBuild(CuTest* tc) {
 	KSI_DataHash *out = NULL;
 	KSI_DataHash *exp = NULL;
 	unsigned char buf[1024];
-	unsigned buf_len;
+	size_t buf_len;
 	int res;
 
 	KSI_ERR_clearErrors(ctx);
@@ -196,7 +196,7 @@ static void testCalChainBuild(CuTest* tc) {
 static void testAggrChainBuilt(CuTest *tc) {
 	int res;
 	unsigned char buf[1024];
-	unsigned buf_len;
+	size_t buf_len;
 	KSI_LIST(KSI_HashChainLink) *chn = NULL;
 	KSI_DataHash *in = NULL;
 	KSI_DataHash *out = NULL;
@@ -254,7 +254,7 @@ static void testAggrChainBuilt(CuTest *tc) {
 static void testAggrChainBuiltWithMetaData(CuTest *tc) {
 	int res;
 	unsigned char buf[1024];
-	unsigned buf_len;
+	size_t buf_len;
 	
 	KSI_LIST(KSI_HashChainLink) *chn = NULL;
 	KSI_DataHash *in = NULL;
